Rename Activity and ViewModel
